[REFACTOR] split packages
- split sources into packages
- remove junk
