+++ src/main/java/sentiments/data/TweetProvider.java
@@ -1,4 +1,6 @@
-package sentiments;
+package sentiments.data;
+
+import sentiments.domain.model.AbstractTweet;
 
 public interface TweetProvider {
