+++ src/main/java/sentiments/domain/model/AbstractTweet.java
@@ -1,4 +1,4 @@
-package sentiments;
+package sentiments.domain.model;
 
 import java.sql.Timestamp;
